+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.1] - 2023-07-26
+
 ### Added
 
 * Security hotspots are now suppressed when they match reviewed server hotspots
@@ -47,6 +49,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   resolved issue retrieval
 * Support for reading `sonar-project.properties`
 
-[unreleased]: https://github.com/Integrated-Application-Development/delphilint/compare/v0.1.0...HEAD
+[unreleased]: https://github.com/Integrated-Application-Development/delphilint/compare/v0.2.1...HEAD
+[0.2.1]: https://github.com/Integrated-Application-Development/delphilint/compare/v0.2.0...v0.2.1
 [0.2.0]: https://github.com/Integrated-Application-Development/delphilint/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/Integrated-Application-Development/delphilint/releases/tag/v0.1.0